[core,component] move Query models to better package
