+++ heroic-core/src/main/java/com/spotify/heroic/http/query/QueryResource.java
@@ -25,6 +25,8 @@ import com.google.common.collect.ImmutableMap;
 import com.spotify.heroic.Query;
 import com.spotify.heroic.QueryManager;
 import com.spotify.heroic.common.JavaxRestFramework;
+import com.spotify.heroic.metric.QueryMetrics;
+import com.spotify.heroic.metric.QueryMetricsResponse;
 import com.spotify.heroic.metric.QueryResult;
 import eu.toolchain.async.AsyncFramework;
 import eu.toolchain.async.AsyncFuture;
