+++ heroic-core/src/test/java/com/spotify/heroic/http/query/QueryMetricsResponseTest.java
@@ -1,5 +1,6 @@
 package com.spotify.heroic.http.query;
 
+import com.spotify.heroic.metric.QueryMetricsResponse;
 import com.spotify.heroic.test.LombokDataTest;
 import org.junit.Test;
